권한설정 및 평점o 테스트x

+++ src/main/java/com/example/block/dto/ReviewResponseDTO.java
@@ -41,7 +41,7 @@ public class ReviewResponseDTO {
         Integer reviewId;
         String userName;
         String prize;
-        float score;
+        Double score;
     }
     @Builder
     @Getter
